requirements.txt
astroid==2.4.2
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
